backend/requirements_linux.txt
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.3.0
black==25.1.0
cffi==1.17.1
click==8.1.8
cryptography==44.0.2
dnspython==2.7.0
dotenv==0.9.9
ecdsa==0.19.1
fastapi==0.115.12
greenlet==3.1.1
h11==0.14.0
idna==3.10
motor==3.7.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.7
psycopg2-binary==2.9.10
pyasn1==0.4.8
pycparser==2.22
pydantic==2.11.2
pydantic_core==2.33.1
pymongo==4.12.0
python-dotenv==1.1.0
python-jose==3.4.0
pytz==2025.2
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.1
typing-inspection==0.4.0
typing_extensions==4.13.1
uvicorn==0.34.0
